basic_slack_backend/for_reference_only_intermediary-service/cardano-ai-slack-bridge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.27.1
httpx==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0